+++ graph/ops_sparse_test.go
@@ -108,7 +108,7 @@ func TestGatherSlices(t *testing.T) {
 			input = IotaFull(g, shapes.Make(dtypes.Float32, 4, 5))
 			start := Const(g, [][]int32{{0}, {1}, {0}}) // Slice from rows 0, 2 and 0 of each example in the batch.
 			sizes := []int{1}                           // Take only one row per start.
-			output = GatherSlices(input, []int{0}, start, sizes)
+			output = GatherSlices(input, []int{0}, start, sizes, true)
 			return
 		}, [][][]float32{{{0, 1, 2, 3, 4}}, {{5, 6, 7, 8, 9}}, {{0, 1, 2, 3, 4}}})
 
@@ -117,7 +117,8 @@ func TestGatherSlices(t *testing.T) {
 			input = IotaFull(g, shapes.Make(dtypes.Float32, 4, 3))
 			start := Const(g, [][]int32{{0}, {1}}) // Slice from rows 0 and 1.
 			sizes := []int{2}                      // Take two rows per start.
-			output = GatherSlices(input, []int{0}, start, sizes)
+			output = GatherSlices(input, []int{0}, start, sizes, true)
+			return
 			return
 		}, [][][]float32{{{0, 1, 2}, {3, 4, 5}}, {{3, 4, 5}, {6, 7, 8}}})
 
@@ -126,7 +127,7 @@ func TestGatherSlices(t *testing.T) {
 			input = IotaFull(g, shapes.Make(dtypes.Float32, 4, 10))
 			start := Const(g, []int32{1, 1}) // Slice in middle of matrix.
 			sizes := []int{2, 3}             // Take a sub-matrix
-			output = GatherSlices(input, []int{0, 1}, start, sizes)
+			output = GatherSlices(input, []int{0, 1}, start, sizes, true)
 			return
 		}, [][]float32{{11, 12, 13}, {21, 22, 23}})
 }